(GH-265) Scaffold tool & related structs
This commit adds the first library file to PRM, scaffolding out:

- The Tool struct
- The ToolExitCode enum
- The ToolConfig struct
- The BinaryConfig struct
- The CommonConfig struct
- The ContainerConfig struct
- The GemConfig struct
- The PuppetConfig struct

These are all public/exported structs/enums so they can be
referenced elsewhere/reused.

This commit also adds some nolint directives on the file to prevent
CI Linting from failing because the scaffolded structs are not used
by any functions yet; as the library is implemented, we will be able
to reenable this lint check.

Resolves puppetlabs/pdkgo#265