(GH-18) Add docker backend status method
Prior to this commit, the docker backend was scaffolded but unimplemented.

This commit adds the new BackendStatus struct, which generically reports
whether a particular backend is available and relevent information about
that backend; in the case of an unavailable backend, it should report
*why* that backend is unavailable (surfacing any error messages) and in
the case of an available backend, useful information.

This commit adds a new DockerClient interface to the docker backend in
order to enable mocking/unit testing. The only function currently used
from the client is the ServerVersion function, consumed in the new Status()
function for the backend.

The Status() function is implemented on the docker backend to call the
ServerVersion() docker client function. If this call errors, then the
docker service/backend is unavailable. The most likely reason for this
is the Docker Engine not running, so special handling is added for that
edge case to handle it in a slightly more friendly way for end users.

If the ServerVersion() docker client function does *not* error it returns
information about Docker which is useful to surface; in this case, the
platform (Docker community/enterprise, desktop/server), the version, and
the API version being used.

Finally, this commit adds unit testing for the new Status() function and
includes mocks for the Docker Client. In this PR the mocks are kept in the
same file as the unit tests; if/when the mocks need to be used in another
test file, it may make sense to move them into a mocks folder.

Resolves #18.