(GH-16) Add `prm get puppet` subcommand; load config on app start
This commit adds the `prm get puppet` subcommand, and hooks it in
to the parent `prm get` command.

Also in this commit is a new global var `prm.RunningConfig` which
will hold the desired running configuration. A new function
`LoadConfig` was added too, which is called from `root#InitConfig`
and loads the configuration from disk (`.prm.yaml`) via Viper to
the `RunningConfig` object.

The `prm get puppet` command is quite basic, but still light on
test coverage. It seems more appropriate to test this via
acceptance tests, so a follow up ticket will add those in.