(GH-15) Add `GenerateDefaultCfg` function
This commit adds the `GenerateDefaultCfg` function to the config
pacakge. This ensures that there is a sensible default set for
Puppet version and Backend upon app initialisation.

Also part of this commit is an enhancement to the `prm set` unit
tests, which set up a temporary viper config and ensure the
expected values were set.