(GH-15) Add `mock.Utils` package for `prm set` tests
Prior to this commit, I was setting up a viper instance within the
unit tests for `prm set`. This was not a great idea and the
inevitable deadlock / resource contention issues hit on the Windows
runner on Github.

This commit adds a mock `Utils` package, which is then used by the
`prm set` unit tests to validate the correct parameters are being
sent to viper.

Also part of this commit, I renamed the `WriteConfig` function to
`SetAndWriteConfig` and enhanced it to both save and write the
config. This was mainly to facilitate the above refactoring to
ensure the code was more testable.