(GH-13) Implement prm install
The install command has been created and all appropriate
help information, arguments and parameters have been added to give it the same install
functionality as that found on pct.

Added the `config_processor.go` with functions to appropriately process
the prm config files after they're extracted from the `tar.gz` tool files.
The `ProcessConfig` method is called from the genericised `install` package which is
located in the pdkgo repository.

Added appropriate acceptance and unit testing along with test data and
mocking modules to supplement the testing.