(GH-11) Add explain command and docs package
This commit updates all of the go packages; this was necessary to pick
up the changes in pdkgo and elsewhere as a prerequisite for this commit.

This commit adds the `md` package which embeds our documentation. A
future commit will be needed to remove the local replace directive; this
was necessary temporarily as the package has not been pushed to main.

This commit adds the explain command to render documentation in the
terminal.

Finally, this commit removes the CHANGELOG and CONTRIBUTING files from
the docs folder and adds frontmatter metadata to the vendored README so
that the explain command has a minimal functionality implemented.

A future commit will need to add documentation to be rendered.

This commit resolves puppetlabs/pdkgo#11.