(GH-54) - Implement a timeout  flag
Flag has been integrated into the prm config and possesses a default value.
Value of the flag is now passed through to `docker.go` where it is correctly implemented.
`status.go` updated so that default timeout value is passed through, needed due to code shared between it and `exec.go`.