(GH-95) Add installing and updating tools docs
Added an installing tools page which documents the usage of the `prm
install` command to install tools. Additionally, this page instructs
users on how to update their tools and list installed tools in a
tabular format.

There is currently no functionality to actually update tools so users
are instructed to install the latest version of a tool. Tool version
selection is a feature that should be added to PRM in a later version.