(MAINT) Add tools to release archives
Prior to this commit, the goreleaser would add all the files in the tools
folder in root, if one exists. Additionally, the release workflow had the
step for cloning the package repository both pointing to PCT templates and
commented out.

This commit sets the release workflow to clone the default PRM tools found
in fantastic-adventure prior to running goreleaser and updates the config
for goreleaser to vendor in those tools.