(GH-341) Add build command
Prior to this commit, there was no provided way to package a tool plugin
for PRM and native tar.gz utilities did not work as expected.

This commit imports the `build` command and library package from the
[pdkgo][1] project to add functionality for building tool packages for
distribution without re-implementing code in this project where it isn't
necessary.

This commit also serves as proof of concept for importing/consuming
other commands from pdkgo, such as `install`, `explain`, and `version`.
In future commits, the code for those commands may be refactored/removed
to reduce copied code and ensure any updates to them only have to be
written and maintained in one place.

[1]: https:github.com/puppetlabs/pdkgo