(GH-7) Add stubbed validate command
This commit adds the validate command to the app. In this initial
implementation it does not run any tools. It can list available tools
but does not filter for validators only. If a tool is specified it will
check to see if the tool is available.

Future work will be required to wire this up for validation
functionality. This commit leaves commented code borrowed from the
implementation of exec which may be useful, but must be addressed prior
to release.