(GH-125) Ensure validate list only returns validators
Prior to this commit, the validate command always returned the full list
of available tools, even those which are not validators. The logic for
listing tools was shared between the exec and validate commands.

This commit extends `prm.List()` to take an additional parameter,
`onlyValidators` as a boolean. If `onlyValidators` is passed as true,
skip caching any tools which are not validators.

This commit updates the references to `prm.List()` to pass false for
`onlyValidators` in the exec command and to pass true in the validate
command. This ensures that exec returns the full list of tools and
validate only returns validators.

Resolves #125