(MAINT) Fix docker fix docker logs
Prior to this commit we were seeing strange behaviour when extracting
logs from a container. The log files written seemed to have unexpected
encoding. It turns out this is because the output of ContainerLogs()
returns a multiplexed stream of the container's stdout and stderr.
Therefore we need to demultiplex the stream before we can decode the
logs and write to our log file.

This commit updates the validate code to use the stdcopy package
provided by the docker sdk to demultiplex the stream.

For reference, see the docs:
https://pkg.go.dev/github.com/moby/moby/client#Client.ContainerLogs