(GH-130) Implement multi-tool validate command
Definitions:
- When referring to the `code directory`, I mean the directory
containing the code which the validation tools are executed against.

This commit adds functionality to run multiple validator tools in series
or parallel. Multiple tools can be specified in groups which are
contained within a file named `validate.yml`. This file should be included
in the base of the code directory. See the example of a `validate.yml`
file at the bottom of this commit message.

The group of tools which the user wishes to validate with can be
selected using the `--groups [group_name]` flag.

The number of tools concurrently validating can be specified using the
`--workerCount [int]` flag. The `--serial` flag will effectively
set the `workerCount=1`.

When validation is complete, logs for each validator can be outputted to
either a log file, or to the terminal. The output location can be
specified using the `--resultsView [string]` flag. It can be set to either
'terminal' or 'file'. Single tool validation (not using `validate.yml
file`) will output to the terminal by default, whereas multi-tool
validation will output to log files. Output files will be stored to
`{code directory}/.prm-validate/`.

Appropriate unit and acceptance tests have been added.

Example `validate.yml` file:
```yaml
groups:
  - id: "ci"
    tools:
      - name: puppetlabs/epp
        args: [--example, args]
      - name: puppetlabs/parser
      - name: puppetlabs/puppet-lint
      - name: puppetlabs/onceover
      - name: puppetlabs/puppet-syntax
      - name: puppetlabs/metadata-json-lint
      - name: puppetlabs/spec_puppet
  - id: "quick_validate"
    tools:
      - name: puppetlabs/epp
      - name: puppetlabs/rubocop
      - name: puppetlabs/parser
      - name: puppetlabs/r10k
```

Note: Worker pool implementation adapted from
https://brandur.org/go-worker-pool.