(MAINT) Better error outputting to terminal
This commit changes how errors are outputted to the terminal.
Previously, when a tool returned a non-success exit code and no stderr,
the terminal would just output an exit code. Now, when tool returns a
non-success exit code and no stderr, the stdout will be outputted to the
terminal. The tool's stdout should contain what validation errors occurred.

This commit also adds a change to how the code directory is
presented in the outputted text. Previously, any usage of a code directory
would have `/code/` prepended onto it as it was mounted to this docker directory.
Now PRM will find and remove any usage of `/code/` in the output text.